backend/pyproject.toml
[project]
name = "ai-orchestrator-backend"
version = "0.1.0"
description = "ai_orchestrator_backend"
authors = [
    {name = "norama",email = "norama.matema@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
ruff = "^0.14.9"
debugpy = "^1.8.19"
pyright = "^1.1.407"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88